requirements.txt
absl-py==2.1.0
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
Box2D==2.3.10
box2d-py==2.3.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chess==1.9.4
click==8.1.8
cloudpickle==3.1.1
decorator==5.2.1
dm-tree==0.1.9
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
gymnasium==1.0.0
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lz4==4.4.3
MarkupSafe==3.0.2
ml_collections==1.0.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
open-spiel==1.5
ormsgpack==1.7.0
packaging==24.2
pandas==2.2.3
pettingzoo==1.24.3
pillow==11.1.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.3.0
pymunk==6.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
ray==2.43.0
referencing==0.36.2
requests==2.32.3
rlcard==1.0.5
rpds-py==0.23.1
scipy==1.15.2
sentry-sdk==2.22.0
setproctitle==1.3.5
Shimmy==2.0.0
six==1.17.0
smmap==5.0.2
SuperSuit==3.9.3
swig==4.3.1
sympy==1.13.1
tensorboardX==2.6.2.2
tensorflow-probability==0.25.0
termcolor==2.5.0
tinyscaler==1.2.8
torch==2.6.0
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.7
wrapt==1.17.2
